[Core] Split module definitions into dedicated header
* Added generated module TDL and header for ModHeader, objModule and module flags
* Updated modules to include the module header where module loading APIs are used
* [Build] Regenerated core metadata and tracked common TDL dependencies for header generation
