+++ docs/xml/modules/classes/xquery.xml
@@ -47,7 +47,9 @@
 if (query.ok()) {
    XPathValue *result;
    if (!query-&gt;evaluate(xml)) {
-      log.msg(&quot;Got: %s&quot;, query-&gt;get&lt;CSTRING&gt;(FID_ResultString));
+      std::string_view str;
+      query-&gt;getResultString(str);
+      log.msg(&quot;Got: %s&quot;, str.data());
    }
 }
 </pre>
