+++ docs/xml/modules/core.xml
@@ -2294,12 +2294,6 @@ if (not OpenDir(path, RDF::FILE|RDF::FOLDER, &amp;info)) {
       <const name="WRITE">The memory is explicitly marked as writeable.</const>
     </constants>
 
-    <constants lookup="MHF" comment="Internal options for requesting function tables from modules.">
-      <const name="DEFAULT">Use structures to group exported functions (Linux, C/C++ standard)</const>
-      <const name="STATIC">Keep the module code in memory</const>
-      <const name="STRUCTURE">Use structures to group exported functions (Linux, C/C++ standard)</const>
-    </constants>
-
     <constants lookup="MSF" comment="Message flags.">
       <const name="NO_DUPLICATE">If the <code>Type</code> parameter matches a message already inside the queue, the new message will not be added and the function will immediately return with <code>ERR::Okay</code>.</const>
       <const name="UPDATE">If <code>Type</code> matches a queued message, update that message's data in place.</const>
