+++ src/core/CMakeLists.txt
@@ -25,7 +25,6 @@ if (KOTUKU_DEFS_AVAILABLE)
    set (CORE_OUTPUTS
       "${DOCS}/modules/core.xml"
       "${DOCS}/modules/classes/metaclass.xml"
-      "${DOCS}/modules/classes/module.xml"
       "${INCLUDE_OUTPUT}/modules/core.h"
       "${INCLUDE_OUTPUT}/system/errors.h"
       "${INCLUDE_OUTPUT}/system/registry.h"
@@ -43,12 +42,12 @@ if (KOTUKU_DEFS_AVAILABLE)
       COMMAND ${CMAKE_COMMAND} -E touch "${CORE_STAMP}"
       DEPENDS
          "${CMAKE_CURRENT_SOURCE_DIR}/defs/core.tdl"
+         "${CMAKE_CURRENT_SOURCE_DIR}/defs/common.tdl"
          "${CMAKE_CURRENT_SOURCE_DIR}/defs/errors.tdl"
          "${CMAKE_CURRENT_SOURCE_DIR}/defs/registry.tdl"
          "${CMAKE_CURRENT_SOURCE_DIR}/../../tools/idl/common.tdl"
          ${FUNCTIONS}
          "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_metaclass.cpp"
-         "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_module.cpp"
       WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
       COMMENT "Producing definition files for the Core library"
       VERBATIM)
@@ -69,6 +68,18 @@ idl_gen ("defs/config.tdl"
    DEPENDS
       "${CMAKE_CURRENT_SOURCE_DIR}/defs/common.tdl"
       "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_config.cpp")
+
+# Module class
+
+idl_gen ("defs/module.tdl"
+   NAME module_defs
+   OUTPUT "${INCLUDE_OUTPUT}/modules/module.h"
+   BYPRODUCTS
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_module_def.c"
+      "${DOCS}/modules/classes/module.xml"
+   DEPENDS
+      "${CMAKE_CURRENT_SOURCE_DIR}/defs/common.tdl"
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_module.cpp")
 
 # Script class
