+++ src/core/defs/module.tdl
@@ -1,0 +1,102 @@
+--$TIRI:Include
+
+header({ module="Core", copyright="Paul Manias © 1996-2026", timestamp=20240611 }, function()
+  c_include('<kotuku/main.h>')
+
+  restrict(function()
+    loadFile('./common.tdl')
+  end)
+
+  c_insert([[
+#ifdef KOTUKU_STATIC
+__export void CloseCore(void);
+__export ERR OpenCore(struct OpenInfo *, struct CoreBase **);
+#else
+__export struct ModHeader ModHeader;
+#endif
+
+#ifdef MOD_NAME
+ #ifdef KOTUKU_STATIC
+  #define KOTUKU_MOD(init,close,open,expunge,test,IDL,Structures) static struct ModHeader ModHeader(init, close, open, expunge, test, IDL, Structures, TOSTRING(MOD_NAME), TOSTRING(MOD_NAMESPACE));
+ #else
+  #define KOTUKU_MOD(init,close,open,expunge,test,IDL,Structures) struct ModHeader ModHeader(init, close, open, expunge, test, IDL, Structures, TOSTRING(MOD_NAME), TOSTRING(MOD_NAMESPACE));
+ #endif
+ #define MOD_PATH ("modules:" TOSTRING(MOD_NAME))
+#else
+ #define MOD_NAME nullptr
+#endif
+
+using ModClose   = void (*)(OBJECTPTR);
+using ModInit    = ERR (*)(OBJECTPTR, struct CoreBase*);
+using ModOpen    = ERR (*)(OBJECTPTR);
+using ModExpunge = ERR (*)(void);
+using ModTest    = void (*)(std::string_view, int *, int *);
+]])
+
+  struct("ModHeader", { comment="Module file header", restrict="c" }, [[
+    int(MHF) Flags                 # Special flags, type of function table wanted from the Core
+    cstr Definitions               # Module definition string, usable by run-time languages such as Tiri
+    fptr(error obj struct(*CoreBase)) Init # A one-off initialisation routine for when the module is first opened.
+    fptr(void obj) Close           # A function that will be called each time the module is closed.
+    fptr(error obj) Open           # A function that will be called each time the module is opened.
+    fptr(error) Expunge            # Reference to an expunge function to terminate the module.
+    fptr(void strview &int &int) Test # A function that can run embedded unit tests in development builds.
+    cstr Name                      # Name of the module
+    cstr Namespace                 # A reserved system-wide namespace for function names.
+  ]],
+  [[
+   typedef const std::vector<std::pair<std::string, StructInfo>> STRUCTS;
+   STRUCTS *StructDefs;
+   class objRootModule *Root;
+
+   ModHeader(ModInit pInit, ModClose pClose, ModOpen pOpen, ModExpunge pExpunge, ModTest pTest,
+      CSTRING pDef, STRUCTS *pStructs, CSTRING pName, CSTRING pNamespace) {
+      Flags         = MHF::DEFAULT;
+      Definitions   = pDef;
+      StructDefs    = pStructs;
+      Init          = pInit;
+      Close         = pClose;
+      Open          = pOpen;
+      Expunge       = pExpunge;
+      Test          = pTest;
+      Name          = pName;
+      Namespace     = pNamespace;
+      Root          = nullptr;
+   }
+  ]])
+
+  methods("Module", "Mod", {
+    { id=1, name="ResolveSymbol" },
+    { id=2, name="Test" }
+  })
+
+  klass("Module", { src="../classes/class_module.cpp", extended=true }, [[
+    cstruct(*Function) FunctionList # Array of functions
+    ptr ModBase                     # Ptr to function jump table
+    obj(RootModule) Root            # Shared details on the module
+    struct(*ModHeader) Header       # For creating in-memory modules only.
+    string Name                     # Name of the module
+    int(MOF) Flags                  # Optional flags
+  ]],
+  nil,
+  [[
+   static ERR load(std::string Name, OBJECTPTR *Module = nullptr, APTR Functions = nullptr) {
+      if (auto module = objModule::create::global(kt::FieldValue(kt::strhash("name"), std::string_view(Name)))) {
+         #ifdef KOTUKU_STATIC
+            if (Module) *Module = module;
+            if (Functions) ((APTR *)Functions)[0] = nullptr;
+            return ERR::Okay;
+         #else
+            APTR functionbase;
+            if (!module->getModBase(functionbase)) {
+               if (Module) *Module = module;
+               if (Functions) ((APTR *)Functions)[0] = functionbase;
+               return ERR::Okay;
+            }
+            else return ERR::GetField;
+         #endif
+      }
+      else return ERR::CreateObject;
+   }
+  ]])
+end)
