+++ src/document/document.cpp
@@ -37,6 +37,7 @@ that is distributed with this package.  Please refer to it for further informati
 #include <kotuku/modules/script.h>
 #include <kotuku/modules/xquery.h>
 #include <kotuku/modules/time.h>
+#include <kotuku/modules/module.h>
 #include <atomic>
 
 using BYTECODE = uint32_t;
