+++ src/image_jpeg/jpeg.cpp
@@ -15,6 +15,7 @@ related to this Package.  The original libjpeg source code can be obtained from 
 #include <kotuku/modules/image.h>
 #include <kotuku/modules/display.h>
 #include <kotuku/modules/filesystem.h>
+#include <kotuku/modules/module.h>
 
 #include "../image/image.h"
