+++ src/xml/xml.cpp
@@ -46,8 +46,9 @@ document access.  XQuery support excludes the following:
 
 #include <kotuku/modules/xml.h>
 #include <kotuku/modules/xquery.h>
-#include <kotuku/strings.hpp>
 #include <kotuku/modules/script.h>
+#include <kotuku/modules/module.h>
+#include <kotuku/strings.hpp>
 #include <array>
 #include <format>
 #include <functional>
